# Make cops rendering much faster