--- a/internal/cops/display/ansi.go
+++ b/internal/cops/display/ansi.go
@@ -5,40 +5,40 @@ import (
 	"strconv"
 )
 
-func renderNoColor(buf []byte, c color.Color) []byte {
+func renderNoColor(buf []byte, c color.RGBA) []byte {
 	return buf
 }
 
-func renderBackgroundColor3(buf []byte, c color.Color) []byte {
+func renderBackgroundColor3(buf []byte, c color.RGBA) []byte {
 	return renderBackgroundColor(buf, Palette3, c)
 }
 
-func renderForegroundColor3(buf []byte, c color.Color) []byte {
+func renderForegroundColor3(buf []byte, c color.RGBA) []byte {
 	return renderForegroundColor(buf, Palette3, c)
 }
 
-func renderBackgroundColor4(buf []byte, c color.Color) []byte {
+func renderBackgroundColor4(buf []byte, c color.RGBA) []byte {
 	return renderBackgroundColor(buf, Palette4, c)
 }
 
-func renderForegroundColor4(buf []byte, c color.Color) []byte {
+func renderForegroundColor4(buf []byte, c color.RGBA) []byte {
 	return renderForegroundColor(buf, Palette4, c)
 }
 
-func renderBackgroundColor8(buf []byte, c color.Color) []byte {
+func renderBackgroundColor8(buf []byte, c color.RGBA) []byte {
 	return renderBackgroundColor(buf, Palette8, c)
 }
 
-func renderForegroundColor8(buf []byte, c color.Color) []byte {
+func renderForegroundColor8(buf []byte, c color.RGBA) []byte {
 	return renderForegroundColor(buf, Palette8, c)
 }
 
-func renderForegroundColor(buf []byte, p color.Palette, c color.Color) []byte {
+func renderForegroundColor(buf []byte, p color.Palette, c color.RGBA) []byte {
 	i := p.Index(c)
 	return renderForegroundColorIndex(buf, i)
 }
 
-func renderBackgroundColor(buf []byte, p color.Palette, c color.Color) []byte {
+func renderBackgroundColor(buf []byte, p color.Palette, c color.RGBA) []byte {
 	i := p.Index(c)
 	return renderBackgroundColorIndex(buf, i)
 }
@@ -77,30 +77,29 @@ func renderBackgroundColorIndex(buf []byte, i int) []byte {
 	return buf
 }
 
-func renderForegroundColor24(buf []byte, c color.Color) []byte {
-	if i, ok := colorIndex[rgba(c)]; ok {
+func renderForegroundColor24(buf []byte, c color.RGBA) []byte {
+	if i, ok := colorIndex[c]; ok {
 		return renderForegroundColorIndex(buf, i)
 	}
 	return renderColor24(buf, "38", c)
 }
 
-func renderBackgroundColor24(buf []byte, c color.Color) []byte {
-	if i, ok := colorIndex[rgba(c)]; ok {
+func renderBackgroundColor24(buf []byte, c color.RGBA) []byte {
+	if i, ok := colorIndex[c]; ok {
 		return renderBackgroundColorIndex(buf, i)
 	}
 	return renderColor24(buf, "48", c)
 }
 
-func renderColor24(buf []byte, code string, c color.Color) []byte {
-	r, g, b, _ := c.RGBA()
+func renderColor24(buf []byte, code string, c color.RGBA) []byte {
 	buf = append(buf, "\033["...)
 	buf = append(buf, code...)
 	buf = append(buf, ";2;"...)
-	buf = append(buf, strconv.Itoa(int(r/256))...)
+	buf = append(buf, strconv.Itoa(int(c.R))...)
 	buf = append(buf, ";"...)
-	buf = append(buf, strconv.Itoa(int(g/256))...)
+	buf = append(buf, strconv.Itoa(int(c.G))...)
 	buf = append(buf, ";"...)
-	buf = append(buf, strconv.Itoa(int(b/256))...)
+	buf = append(buf, strconv.Itoa(int(c.B))...)
 	buf = append(buf, "m"...)
 	return buf
 }