Nearly 5x faster! Now with 0 allocs!