@jcorbin 2017-12-09
Before:
```
Benchmark_displayDemo/4-4          50000             31468 ns/op             816 B/op        200 allocs/op
Benchmark_displayDemo/8-4          10000            125054 ns/op            3334 B/op        830 allocs/op
Benchmark_displayDemo/16-4                  3000            483670 ns/op           13335 B/op       3327 allocs/op
Benchmark_displayDemo/32-4                  1000           1953902 ns/op           53385 B/op      13337 allocs/op
Benchmark_displayDemo/64-4                   200           7905467 ns/op          213224 B/op      53280 allocs/op
Benchmark_displayDemo/128-4                   50          31240162 ns/op          853395 B/op     213263 allocs/op
Benchmark_displayDemo/256-4                   10         120686210 ns/op         3415774 B/op     853734 allocs/op
```

After:
```
Benchmark_displayDemo/4-4         100000             12235 ns/op              16 B/op          1 allocs/op
Benchmark_displayDemo/8-4          30000             46893 ns/op              16 B/op          1 allocs/op
Benchmark_displayDemo/16-4                 10000            184436 ns/op              19 B/op          1 allocs/op
Benchmark_displayDemo/32-4                  2000            734434 ns/op              24 B/op          1 allocs/op
Benchmark_displayDemo/64-4                   500           2944757 ns/op              32 B/op          1 allocs/op
Benchmark_displayDemo/128-4                  100          11762381 ns/op              97 B/op          1 allocs/op
Benchmark_displayDemo/256-4                   30          47267835 ns/op            1980 B/op          1 allocs/op
```

That "1 alloc" is in the benchmark itself.