@jcorbin 2017-12-09
Note however that we're still talking about ~11ms/op for a "reasonable" 128x128 grid; here's the top:
```
Showing top 10 nodes out of 37
      flat  flat%   sum%        cum   cum%
    1290ms 11.44% 11.44%     9830ms 87.15%  github.com/borkshop/bork/internal/cops/display.RenderOver /home/jcorbin/golang/src/github.com/borkshop/bork/internal/cops/display/display.go
    1210ms 10.73% 22.16%     1460ms 12.94%  strconv.formatBits /usr/lib/go-1.9/src/strconv/itoa.go
     840ms  7.45% 29.61%      840ms  7.45%  image.(*RGBA).RGBAAt /usr/lib/go-1.9/src/image/image.go
     800ms  7.09% 36.70%     1910ms 16.93%  github.com/borkshop/bork/internal/cops/display.(*Display).RGBAAt /home/jcorbin/golang/src/github.com/borkshop/bork/internal/cops/display/display.go
     710ms  6.29% 43.00%      960ms  8.51%  runtime.mapaccess2_fast32 /usr/lib/go-1.9/src/runtime/hashmap_fast.go
     700ms  6.21% 49.20%     3170ms 28.10%  github.com/borkshop/bork/internal/cops/display.renderColor24 /home/jcorbin/golang/src/github.com/borkshop/bork/internal/cops/display/ansi.go
     690ms  6.12% 55.32%     2180ms 19.33%  strconv.AppendInt /usr/lib/go-1.9/src/strconv/itoa.go
     670ms  5.94% 61.26%      670ms  5.94%  runtime.memmove /usr/lib/go-1.9/src/runtime/memmove_amd64.s
     640ms  5.67% 66.93%      670ms  5.94%  github.com/borkshop/bork/internal/cops/display.Cursor.Go /home/jcorbin/golang/src/github.com/borkshop/bork/internal/cops/display/cursor.go
     530ms  4.70% 71.63%     5650ms 50.09%  github.com/borkshop/bork/internal/cops/display.(*model).RenderRGBA <autogenerated>
```
Basically our own rendering function is the chief part now, followed by all that ansi formatting.