@jcorbin 2017-12-09
Hmm, turns out I wasn't doing the requisite front/back swap, but RenderOver was never updating the back buffer, so it was a wash anyhow... fixed.