@jcorbin 2017-12-09
Added another ~40-50% speedup by string pre-computation...