@jcorbin 2017-12-09
Added another ~50% gain on prior by ruthlessly indexing rather than do all that x/y => index conversion.