@jcorbin 2017-12-09
ret-conned the under-update correction to fix the bench since inception, to bring you this comparison:
```
name                old time/op    new time/op    delta
_displayDemo/4-4      26.2µs ± 0%     4.8µs ± 0%   ~     (p=1.000 n=1+1)
_displayDemo/8-4       103µs ± 0%      16µs ± 0%   ~     (p=1.000 n=1+1)
_displayDemo/16-4      400µs ± 0%      63µs ± 0%   ~     (p=1.000 n=1+1)
_displayDemo/32-4     1.61ms ± 0%    0.25ms ± 0%   ~     (p=1.000 n=1+1)
_displayDemo/64-4     6.60ms ± 0%    0.98ms ± 0%   ~     (p=1.000 n=1+1)
_displayDemo/128-4    26.3ms ± 0%     4.0ms ± 0%   ~     (p=1.000 n=1+1)
_displayDemo/256-4     105ms ± 0%      16ms ± 0%   ~     (p=1.000 n=1+1)
```

So we`re near 5-6x faster on the whole so far.