@jcorbin 2017-12-09
However the other elephant in the perf room now, is the choice to use strings rather than limit to single rune; i.e. `Display.rgbaati` is now the 2nd highest cpu consumer, and it's all in that first "copy out of the textile" line... suspect will need to do something even more ruthless like a symbol table or a static `[2]rune` to move past this one...