@jcorbin 2017-12-09
Got another ~10-15% by skipping the `Model` interface.