@jcorbin 2017-12-09
Added back sub-display rendering, by using intersection; it may have only cost ~1ms (or a rounding difference) on the worst case bench; verified that braillewave demo looks ok locally; wish we had more tests, especially oriented at full-screen rendering.