@jcorbin 2017-12-09
I left some other ideas that turned out to not matter on https://github.com/borkshop/bork/compare/cops_moar_perf ; I'm especially fond off the `Cursor.Go` breakup even if it's not (yet) material to perf.